build: move supporting js files into `src`
